+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2025-10-19
+
+### Added
+
+- Style options for text tooltips.
+- Recovery compass tooltips.
+- Clock tooltips.
+- `en_ud` translation.
+- Tooltips for light level.
+
+### Changed
+- Non-attribute icons are now placed in the Inline Tooltips namespace.
+- Non-attribute icons are no longer hidden when Shift is held.
+
+### Fixed
+- Mipmapping issues.
+
 ## [1.2.0] - 2025-10-16
 
 Initial release.